Moved Swarm Connect from node API to plum pork
